// src/Types.ts
export type FormChild = {
    [x: string]: any;
    record_display_name?: any;
    name?: string[];
    label?: string;
    colspan?: string;
    not_empty?: string;
    is_web_compatible?: string;
    is_mobile_compatible?: string;
    helper: string;
    method: string;
    field_collection?: string;
    allow_copy_to_clipboard?: string;
    template: string;
    packaged_with_module: string;
    mw?: boolean;
    children?: FormChild[];
    rules_grid?: string;
    roles_grid?: string;
    disable_autofill_while_cloning?: string;
    allow_user_to_specify_user_level_default?: string;
    business_key?: string;
    __name?: string;
    inline_search?: string;
    select?: string;
    open?: string;
    show_model?: string;
    show_link?: string;
    concat_text_value?: string;
    postfix_label?: string;
    prefix_id?: string;
    value?: string;
    value_model?: string;
    __value?: string;
    on_change_reload_form?: string;
    permission?: string;
    filter_by_fields?: string;
    __filter_by_fields?: string;
    help?: string;
    tooltip?: string;
    editor?: string;
    ai_provider_id_model?: string;
    auto_resize?: string;
    allow_commenting_on_this_field?: string;
    rules?: Rule[];
    load_menu?: string;
    listview?: string;
    listview_model?: string;
    __listview?: string;
    pfooter?: string;
    value_1?: string;
    value_2?: string;
    is_first_level_filter?: string;
    pheader?: string;
    hide_listview_link?: string;
    show_in_draft_mode?: string;
    disable_search?: string;
    show_opened?: string;
    append_main_model_filter?: string;
    module?: string;
    controller?: string;
    action?: string;
    web_compatible?: string;
    mobile_compatible?: string;
    sub_form_model_class?: string;
    sub_form_name?: string;
    overwrite_sub_form_model?: string;
    show_collapsed?: string;
    description?: string;
    placeholder?: string;
    primary_key?: string;
    q?: ListQuery;
    string?: string;
    href?: string;
    model?: string;
    text?: string;
    lookup?: string;
    lookup_class?: string;
    roles?: Record<string, RolePermission>;
    add_button_bar?: string;
    allow_row_create?: string;
    allow_row_delete?: string;
    js_statements?: string;
    js_formula?: string;
    show_labels?: string;
    min?: string; // In GridTemplate, min refers to the minimum rows user need to input. In DateTimeTemplate, min refers to the minimum date that users are allowed to choose
    max?: string;
};

type Rule = {
    if_condition: string;
    permission: string;
    id: string;
    deleted: string;
    is_active: string;
};
type ListQuery = {
    pfooter?: string;
    pheader?: string;
    method?: string;
    fields?: string[];
    parent_id?: string;
    active_level?: string;
    where?: {
        [key: string]: string;
    };
    parent_where?: {
        [key: string]: string;
    };
    related_model_class?: string;
    show_has_many_records?: string;
};

type RolePermission = {
    __role_id: string;
    _is_active: string;
    is_active: string;
    permission: string;
    role_id: string;
    role_id_model: string;
};